Review of "Bring consistency among sizes"
Issue #928

+++ address.c
@@ -848,7 +848,7 @@ bool mutt_addr_valid_msgid(const char *msgid)
     return false;
 
   /* TODO: complete parser */
-  for (unsigned int i = 0; i < l; i++)
+  for (size_t i = 0; i < l; i++)
     if ((unsigned char) msgid[i] > 127)
       return false;
