+++ curs_main.c
@@ -700,10 +700,10 @@ int index_color(int index_no)
  */
 void mutt_draw_statusline(int cols, const char *buf, int buflen)
 {
-  int i = 0;
-  int offset = 0;
+  size_t i = 0;
+  size_t offset = 0;
   bool found = false;
-  int chunks = 0;
+  size_t chunks = 0;
   size_t len = 0;
 
   struct Syntax
@@ -785,7 +785,7 @@ void mutt_draw_statusline(int cols, const char *buf, int buflen)
     if (len <= syntax[i].last)
       goto dsl_finish; /* no more room */
 
-    int next;
+    size_t next;
     if ((i + 1) == chunks)
     {
       next = len;
